src/requirements.txt
torch
numpy
pillow
scipy
matplotlib
scikit-image
scikit-learn
lpips
pyyaml
kagglehub